+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## 37.0-SNAPSHOT - unreleased
 
+### 🎁 New Features
+
+* Enable underlying support for Spring Boot Actuator Endpoints.
+     * Expose these endpoints in your application via configuration in your application.groovy
+        file, e.g. 'management.endpoints.web.exposure.include = "health,info,metrics"`
+     * See https://docs.spring.io/spring-boot/3.5/reference/actuator/endpoints for more info on
+       the available endpoints and how to configure and use them.
+
+
 ## 36.0.0 - 2026-01-27
 
 ### 💥 Breaking Changes (upgrade difficulty: 🟢 LOW, excepting multi-instance apps w/websockets)